[ALPS07779043] battery: fix type errors
Fix the type errors reported by clang-r475365 (clang 16.0.1)

Errors:
drivers/power/supply/mtk_charger.c:537:3: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_charger.c:1840:5: error: format specifies type long but the argument has type int [-Werror,-Wformat]
drivers/power/supply/mtk_charger.c:3356:36: error: format specifies type int but the argument has type unsigned long [-Werror,-Wformat]
drivers/power/supply/mtk_pd_hal.c:512:46: error: format specifies type unsigned long but the argument has type u32 * (aka unsigned int *) [-Werror,-Wformat]
drivers/power/supply/mtk_pd_hal.c:531:46: error: format specifies type unsigned long but the argument has type u32 (aka unsigned int) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:1205:31: error: format specifies type int but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:2929:32: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:3226:32: error: format specifies type int but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:3228:25: error: format specifies type int but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:3230:3: error: format specifies type int but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:3241:28: error: format specifies type int but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:3250:39: error: format specifies type int but the argument has type struct timespec64 [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:4033:36: error: format specifies type char * but the argument has type int * [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:4047:36: error: format specifies type char * but the argument has type int * [-Werror,-Wformat]
drivers/power/supply/mtk_battery.c:4060:36: error: format specifies type char * but the argument has type int * [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:274:6: error: format specifies type int but the argument has type unsigned long [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:276:6: error: format specifies type int but the argument has type unsigned long [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:1662:40: error: format specifies type int but the argument has type struct timespec64 [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:2372:45: error: format specifies type int but the argument has type size_t (aka unsigned long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:2428:14: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:2429:14: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:3516:33: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:4866:3: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mtk_battery_daemon.c:4883:4: error: format specifies type long but the argument has type time64_t (aka long long) [-Werror,-Wformat]
drivers/power/supply/mt6375-gauge.c:284:56: error: format specifies type int but the argument has type unsigned long [-Werror,-Wformat]
drivers/power/supply/k6985v1_64_alpha/bq28z610.c:396:54: error: format specifies type int but the argument has type union power_supply_propval [-Werror,-Wformat]
drivers/power/supply/k6985v1_64_alpha/bq28z610.c:396:67: error: format specifies type int but the argument has type union power_supply_propval [-Werror,-Wformat]
drivers/power/supply/k6985v1_64_alpha/bq28z610.c:396:77: error: format specifies type int but the argument has type union power_supply_propval [-Werror,-Wformat]

MTK-Commit-Id: 19a4bf08547341156605159d6773b79dc9b5d9f7

CR-Id: ALPS07779043
Feature: Kernel SI Operation

+++ drivers/power/supply/mtk_charger.c
@@ -533,7 +533,7 @@ static void mtk_charger_start_timer(struct mtk_charger *info)
 	info->endtime = end_time;
 	ktime = ktime_set(info->endtime.tv_sec, info->endtime.tv_nsec);
 
-	chr_err("%s: alarm timer start:%d, %ld %ld\n", __func__, ret,
+	chr_err("%s: alarm timer start:%d, %lld %ld\n", __func__, ret,
 		info->endtime.tv_sec, info->endtime.tv_nsec);
 	alarm_start(&info->charger_timer, ktime);
 }
@@ -1837,7 +1837,7 @@ static ssize_t sc_ibat_limit_store(
 		if (val < 0) {
 			chr_err(
 				"[smartcharging ibat limit] val is %ld ??\n",
-				(int)val);
+				val);
 			val = 0;
 		}
 
@@ -3353,7 +3353,7 @@ int notify_adapter_event(struct notifier_block *notifier,
 int chg_alg_event(struct notifier_block *notifier,
 			unsigned long event, void *data)
 {
-	chr_err("%s: evt:%d\n", __func__, event);
+	chr_err("%s: evt:%lu\n", __func__, event);
 
 	return NOTIFY_DONE;
 }
